Address PR review feedback v2
