+++ charts/gardener/resource-manager/charts/runtime/templates/configmap.yaml
@@ -189,3 +189,8 @@ data:
         {{- end }}
       tokenInvalidator:
         enabled: {{ .Values.global.config.webhooks.tokenInvalidator.enabled }}
+      nodeAgentAuthorizer:
+        enabled: {{ .Values.global.config.webhooks.nodeAgentAuthorizer.enabled }}
+        {{- if .Values.global.config.webhooks.nodeAgentAuthorizer.machineNamespace }}
+        machineNamespace: {{ .Values.global.config.webhooks.nodeAgentAuthorizer.machineNamespace }}
+        {{- end }}
