+++ pkg/gardenlet/operation/botanist/botanist.go
@@ -243,8 +243,8 @@ func New(ctx context.Context, o *operation.Operation) (*Botanist, error) {
 // NodeAgentAuthorizerWebhookReady checks if gardener-resource-manager is running, so that node-agent-authorizer webhook is accessible.
 func (b *Botanist) NodeAgentAuthorizerWebhookReady(ctx context.Context) (bool, error) {
 	// The reconcile flow deploys the kube-apiserver of the shoot before the gardener-resource-manager (it has to be
-	// this way, otherwise the Gardener components cannot start). However, GRM serves an authorization webhook for the
-	// SeedAuthorizer feature. We can only configure kube-apiserver to consult this webhook when GRM runs, obviously.
+	// this way, otherwise the shoot components cannot start). However, GRM serves an authorization webhook for the
+	// NodeAgentAuthorizer feature. We can only configure kube-apiserver to consult this webhook when GRM runs, obviously.
 	// This is not possible in the initial kube-apiserver deployment (due to above order).
 	// Hence, we have to deploy kube-apserver a second time - this time with the NodeAgentAuthorizer feature getting enabled.
 	// From then on, all subsequent reconciliations can always enable it and only one deployment is needed.
