+++ skaffold-operator.yaml
@@ -399,7 +399,7 @@ build:
             - pkg/utils/validation/kubernetes/core
             - pkg/utils/validation/kubernetesversion
             - pkg/utils/version
-            - pkg/utils/webhook/authorizer
+            - pkg/webhook/authorizer
             - third_party/gopkg.in/yaml.v2
             - VERSION
         ldflags:
@@ -885,7 +885,7 @@ build:
             - pkg/utils/timewindow
             - pkg/utils/validation/kubernetesversion
             - pkg/utils/version
-            - pkg/utils/webhook/authorizer
+            - pkg/webhook/authorizer
             - third_party/gopkg.in/yaml.v2
             - VERSION
         ldflags:
