+++ skaffold.yaml
@@ -503,7 +503,7 @@ build:
             - pkg/utils/timewindow
             - pkg/utils/validation/kubernetesversion
             - pkg/utils/version
-            - pkg/utils/webhook/authorizer
+            - pkg/webhook/authorizer
             - third_party/gopkg.in/yaml.v2
             - VERSION
         ldflags:
@@ -1247,8 +1247,8 @@ build:
             - pkg/utils/istio
             - pkg/utils/kubernetes
             - pkg/utils/kubernetes/bootstraptoken
+            - pkg/utils/kubernetes/certificatesigningrequest
             - pkg/utils/kubernetes/client
-            - pkg/utils/kubernetes/csr
             - pkg/utils/kubernetes/health
             - pkg/utils/kubernetes/health/checker
             - pkg/utils/kubernetes/unstructured
@@ -1398,7 +1398,7 @@ build:
             - pkg/utils/validation/kubernetes/core
             - pkg/utils/validation/kubernetesversion
             - pkg/utils/version
-            - pkg/utils/webhook/authorizer
+            - pkg/webhook/authorizer
             - third_party/gopkg.in/yaml.v2
             - VERSION
         ldflags:
@@ -1500,7 +1500,7 @@ build:
             - pkg/utils/gardener
             - pkg/utils/imagevector
             - pkg/utils/kubernetes
-            - pkg/utils/kubernetes/csr
+            - pkg/utils/kubernetes/certificatesigningrequest
             - pkg/utils/kubernetes/health
             - pkg/utils/managedresources
             - pkg/utils/managedresources/builder
